move resolveMetadata into async Metadata component
